any bitlen

+++ pages/src/lib.rs
@@ -8,7 +8,7 @@ pub struct JsSpan {
 }
 
 #[wasm_bindgen]
-pub fn evaluate(s: &str) -> Result<f32, JsSpan> {
+pub fn evaluate(s: &str) -> Result<f64, JsSpan> {
     Ok(to_nodes(s).map_err(|s| JsSpan {
         start: s.start,
         end: s.end,
